# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: School Bus Tracking API
  version: 0.7.0
  description: |
    API for the School Bus Tracking admin web and the Android Driver and Parent apps.

    ## Conventions
    - Base path for business endpoints: `/api/v1`. JSON fields are `snake_case`.
    - Timestamps are RFC 3339 UTC (`2026-10-07T02:15:00Z`); calendar dates are `YYYY-MM-DD`.
    - Mobile numbers are accepted as `9876543210`, `+91 98765 43210`, etc. and always returned as `+919876543210`.
    - Success responses wrap the payload: `{"data": ...}`. Lists add
      `"meta": {"page", "page_size", "total"}` and accept `?page=1&page_size=20` (max 100).
    - Errors: `{"error": {"code", "message", "fields"}}`. Branch on `code`, show `message`.
    - Every response carries an `X-Request-ID` header; include it when reporting a bug.

    ## Authentication
    - **Driver and Parent apps:** `POST /auth/otp/send` then `POST /auth/otp/verify` with `app` = `driver` or `parent`.
      On the dev environment every code is `123456` (no SMS is sent).
    - **Admin web:** `POST /auth/login` (email + password). If the response has `mfa_required: true`,
      call `POST /auth/login/2fa` with the `mfa_token` and the authenticator code.
    - Send `Authorization: Bearer <access_token>` on every other call.
    - Access tokens last `expires_in` seconds (15 min). On `401`, call `POST /auth/refresh` once with the
...
      latest one. If refresh fails, send the user to login.
...

...

    ### Driver app: sending GPS
...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/schools/{schoolID}/routes/{routeID}:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
      - { $ref: "#/components/parameters/RouteID" }
    get:
      tags: [Routes & stops]
      summary: Get a route with its stops in order
      operationId: getRoute
      responses:
        "200":
          description: The route; `stops` sorted by `sequence`.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Route" } } }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Routes & stops]
      summary: Update a route
      operationId: updateRoute
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RouteInput" }
      responses:
        "200":
          description: Updated (without stops).
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Route" } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/v1/schools/{schoolID}/students:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
    get:
      tags: [Students]
      summary: List students (Transport Manager may view)
      description: Each student includes the current `assignment`. Transport Managers do not receive `notes`.
      operationId: listStudents
      parameters:
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PageSize" }
        - { name: q, in: query, description: Search name or admission number, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [active, inactive] } }
        - { name: transport_status, in: query, schema: { type: string, enum: [uses_transport, not_using] } }
        - { name: class, in: query, schema: { type: string } }
        - { name: section, in: query, schema: { type: string } }
        - { name: route_id, in: query, schema: { type: string, format: uuid } }
        - { name: assigned, in: query, description: Only students with (yes) or without (no) a route, schema: { type: string, enum: ["yes", "no"] } }
      responses:
        "200":
          description: Students, by class, section, name.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Student" } }
                  meta: { $ref: "#/components/schemas/PageMeta" }
    post:
      tags: [Students]
      summary: Create a student (Super Admin, School Admin)
      operationId: createStudent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StudentInput" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Student" } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/v1/schools/{schoolID}/students/{studentID}:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
      - { $ref: "#/components/parameters/StudentID" }
    get:
      tags: [Students]
      summary: Get a student with linked parents
      operationId: getStudent
      responses:
        "200":
          description: The student.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Student" } } }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Students]
      summary: Update a student
      description: Setting `transport_status` to `not_using` ends the route assignment.
      operationId: updateStudent
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StudentInput" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Student" } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/schools/{schoolID}/students/{studentID}/status:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
      - { $ref: "#/components/parameters/StudentID" }
    patch:
      tags: [Students]
      summary: Activate or deactivate a student
      operationId: setStudentStatus
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties: { status: { type: string, enum: [active, inactive] } }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Student" } } }
  /api/v1/schools/{schoolID}/students/{studentID}/assignment:
    parameters:
      - { $ref: "#/components/parameters/SchoolID" }
      - { $ref: "#/components/parameters/StudentID" }
    put:
      tags: [Students]
      summary: Assign route and stops
      description: |
        Stops must belong to the route. A pickup stop is required if the route runs Morning Pickup and must be empty
        otherwise; the same for the drop stop and Evening Drop. The previous assignment is kept as history.
      operationId: assignStudent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [route_id]
              properties:
                route_id: { type: string, format: uuid }
                pickup_stop_id: { type: [string, "null"], format: uuid }
                drop_stop_id: { type: [string, "null"], format: uuid }
      responses:
        "200":
          description: The student with the new assignment.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: "#/components/schemas/Student" } } }
        "400": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      tags: [Students]
      summary: Remove from route
      operationId: unassignStudent
      responses:
        "204": { description: Removed (kept as history). }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SchoolID: { name: schoolID, in: path, required: true, schema: { type: string, format: uuid } }
    RouteID: { name: routeID, in: path, required: true, schema: { type: string, format: uuid } }
    StudentID: { name: studentID, in: path, required: true, schema: { type: string, format: uuid } }
...
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { name: page_size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
  schemas:
...
    StopRef:
      type: object
      required: [id, name, sequence, latitude, longitude, pickup_time, drop_time]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        sequence: { type: integer }
        latitude: { type: number }
        longitude: { type: number }
        pickup_time: { type: [string, "null"], description: "HH:MM, Morning Pickup" }
        drop_time: { type: [string, "null"], description: "HH:MM, Evening Drop" }
    StudentAssignment:
      type: object
      required: [id, route_id, route_code, route_name, pickup_stop, drop_stop, assigned_at, ended_at]
      properties:
        id: { type: string, format: uuid }
        route_id: { type: string, format: uuid }
        route_code: { type: string }
        route_name: { type: string }
        pickup_stop:
          oneOf: [{ $ref: "#/components/schemas/StopRef" }, { type: "null" }]
          description: Null if the route does not run Morning Pickup.
        drop_stop:
          oneOf: [{ $ref: "#/components/schemas/StopRef" }, { type: "null" }]
          description: Null if the route does not run Evening Drop.
        assigned_by_name: { type: [string, "null"] }
        assigned_at: { type: string, format: date-time }
        ended_at: { type: [string, "null"], format: date-time, description: Null for the current assignment. }
...
    StudentInput:
      type: object
      required: [admission_no, name]
      properties:
        admission_no: { type: string, description: Unique per school; stored upper-case. }
        name: { type: string }
        class: { type: string, example: "5" }
        section: { type: string, example: A }
        notes: { type: string, description: Admin-only. }
        transport_status: { type: string, enum: [uses_transport, not_using], default: uses_transport }
    Student:
      type: object
      required: [id, school_id, admission_no, name, class, section, status, transport_status, assignment, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        school_id: { type: string, format: uuid }
        admission_no: { type: string }
        name: { type: string }
        class: { type: string }
        section: { type: string }
        notes: { type: string, description: Omitted for Transport Managers. }
        status: { type: string, enum: [active, inactive] }
        transport_status: { type: string, enum: [uses_transport, not_using] }
        assignment:
          oneOf: [{ $ref: "#/components/schemas/StudentAssignment" }, { type: "null" }]
        parents:
          type: array
          description: Only on GET of a single student.
          items:
            type: object
            properties:
              parent_id: { type: string, format: uuid }
              name: { type: string }
              mobile: { type: string }
              relationship: { type: string, enum: [father, mother, guardian] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    RouteInput:
      type: object
      required: [name, code]
      properties:
        name: { type: string }
        code: { type: string, example: RS-01, description: Unique per school; stored upper-case. }
        start_point: { type: string, example: School }
        description: { type: string }
        supports_pickup: { type: boolean, default: true, description: Used for Morning Pickup trips. }
        supports_drop: { type: boolean, default: true, description: Used for Evening Drop trips. }
    Route:
      allOf:
        - $ref: "#/components/schemas/RouteInput"
        - type: object
          required: [id, school_id, status, stop_count, created_at, updated_at]
          properties:
            id: { type: string, format: uuid }
            school_id: { type: string, format: uuid }
            status: { type: string, enum: [active, inactive] }
            stop_count: { type: integer }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
            stops:
              type: array
              items: { $ref: "#/components/schemas/Stop" }
              description: Only on GET of a single route, ordered by sequence.
    StopInput:
      type: object
      required: [name, latitude, longitude]
      properties:
        name: { type: string, example: Gandhipuram }
        landmark: { type: string }
        latitude: { type: number, minimum: -90, maximum: 90, example: 11.0168 }
        longitude: { type: number, minimum: -180, maximum: 180, example: 76.9663 }
        pickup_time:
          type: [string, "null"]
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          example: "07:10"
          description: Scheduled Morning Pickup time (HH:MM).
        drop_time:
          type: [string, "null"]
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          example: "16:40"
          description: Scheduled Evening Drop time (HH:MM).
        geofence_radius_m: { type: integer, minimum: 25, maximum: 1000, default: 100, description: Radius for Reached detection. }
    Stop:
      allOf:
        - $ref: "#/components/schemas/StopInput"
        - type: object
          required: [id, route_id, sequence, created_at, updated_at]
          properties:
            id: { type: string, format: uuid }
            route_id: { type: string, format: uuid }
            sequence: { type: integer, minimum: 1, description: "Travel order, 1..N without gaps." }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            fields:
              type: object
              additionalProperties: { type: string }
    PageMeta:
      type: object
      required: [page, page_size, total]
      properties:
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
...
  responses:
...
    ValidationFailed:
      description: Invalid input.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: { code: validation_failed, message: One or more fields are invalid., fields: { mobile: must be a valid 10-digit mobile number } } }
...
    NotFound:
      description: Missing, or belongs to another school.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Duplicate value.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: { code: conflict, message: A record with this email already exists., fields: { email: already exists } } }
...